Add chans.Bridge example
Fixes devlights/gomy#115
