+++ chans/bridge_test.go
@@ -1,0 +1,1 @@
+package chans_test
